Search by scanned ISBN barcode via Barcode Scanner app (free)
It mangles the ISBN to get it to work (chops off first 3 digits and last single)
I'm not really sure if that's right, but it seems to work pretty well,
and I don't want to use thingISBN partly because of the 1000/day usage limit
and partly because people seem to like the ability for this app to work
offline.

Also just to note this is a completed feature branch.